Refactor FiberId (#5881)
* initial work

* fix typo

* fix tests

* fix js specific issue

* tailrec

* rename

* address review comments

* preserve information on interruptors

+++ test/shared/src/main/scala/zio/test/GenZIO.scala
@@ -30,7 +30,7 @@ trait GenZIO {
     val failure           = e.map(Cause.fail(_))
     val die               = t.map(Cause.die(_))
     val empty             = Gen.const(Cause.empty)
-    val interrupt         = Gen.long.zipWith(Gen.long)((l, r) => Cause.interrupt(FiberId(l, r)))
+    val interrupt         = Gen.int.zipWith(Gen.int)((l, r) => Cause.interrupt(FiberId(l, r)))
     def stackless(n: Int) = Gen.suspend(causesN(n - 1).flatMap(c => Gen.elements(Cause.stack(c), Cause.stackless(c))))
 
     def sequential(n: Int) = Gen.suspend {
